refactor: process translation sync through jobs
